Added user registration, logout. Added participants to the rooms and CRUD functionality for comments in rooms.

+++ base/views.py
@@ -1,9 +1,12 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
 from django.db.models import Q
+from django.contrib.auth.decorators import login_required
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
-from .models import Room, Topic
+from .models import Room, Topic, Message
 from .forms import RoomForm
 
 # Create your views here.
@@ -15,6 +18,11 @@ from .forms import RoomForm
 # ]
 
 def loginpage(request):
+    page = 'login'
+
+    if (request.user.is_authenticated):
+        return redirect('home')
+    
     if (request.method == 'POST'):
         username = request.POST.get('username')
         password = request.POST.get('password')
@@ -30,10 +38,33 @@ def loginpage(request):
             login(request, user)
             return redirect('home')
         else:
-            messages.error(request, 'Username or password do not match.')
+            messages.error(request, 'Username or password do not')
 
-    context = {}
+    context = {'page': page}
     return render(request,'base/login_register.html',context)
+
+def logoutuser(request):
+    logout(request)
+    return redirect('home')
+
+def registeruser(request):
+    page = 'register'
+    form = UserCreationForm()
+
+    if (request.method == 'POST'):
+        form = UserCreationForm(request.POST)
+        if (form.is_valid):
+            user = form.save(commit = False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'An error occured during registration.')
+
+
+    context = {'form' : form}
+    return render(request, 'base/login_register.html',context)
 
 
 def home(request):
@@ -50,22 +81,40 @@ def home(request):
 
 def room(request,pk):
     room = Room.objects.get(id = pk)
-    context = {'room':room}
-    return render(request,'base/room.html',context)
+    participants = room.participants.all()
 
+    if (request.method == 'POST'):
+        message = Message.objects.create(
+            user = request.user,
+            room = room,
+            body = request.POST.get('body'),
+        )
+        room.participants.add(request.user)
+        return redirect('room', pk = room.id)
+
+
+    room_messages = room.message_set.all().order_by('-created')
+    context = {'room' : room, 'room_messages' : room_messages, 'participants' : participants}
+    return render(request,'base/room.html', context)
+
+@login_required(login_url = 'login')
 def createroom(request):
     form = RoomForm()
     if (request.method == 'POST'):
-        form = RoomForm(request.POST,instance=room)
+        form = RoomForm(request.POST)
         if form.is_valid():
             form.save()
             return redirect('home')
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url = 'login')
 def updateroom(request,pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if (request.user != room.host):
+        return HttpResponse("You are not the room host.")
 
     if (request.method == 'POST'):
         form = RoomForm(request.POST,instance=room)
@@ -75,9 +124,26 @@ def updateroom(request,pk):
     context = {'form':form}
     return render(request,'base/room_form.html',context)
 
+@login_required(login_url = 'login')
 def deleteroom(request,pk):
     room = Room.objects.get(id=pk)
+
+    if (request.user != room.host):
+        return HttpResponse("You are not the room host.")
+    
     if (request.method == 'POST'):
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+@login_required(login_url = 'login')
+def deletemessage(request,pk):
+    message = Message.objects.get(id=pk)
+
+    if (request.user != message.user):
+        return HttpResponse("You are not the room host.")
+    
+    if (request.method == 'POST'):
+        message.delete()
+        return redirect('room', pk=message.room.pk)
+    return render(request,'base/delete.html',{'obj':message})
